# Optimize ABLForcingAlgorithm for precursor simulations