--- a/src/CMakeLists.txt
+++ b/src/CMakeLists.txt
@@ -1,5 +1,4 @@
 add_sources(GlobalSourceList
-   ABLForcingAlgorithm.C
    ABLProfileFunction.C
    ActuatorLinePointDrag.C
    Algorithm.C