- Move ABLForcingAlgorithm to `wind_energy` directory

- Convert ABLForcingAlgorithm to use the ABL Statistics object that already
  contains the spatially averaged velocities and temperatures as a function of
  height to determine driving pressure gradient

- Update ablNeutralEdge regression test to use the new ABL forcing class.
  Change in order of operations caused diffs of O(1e-10), updated gold files
  accepting diffs. Test runtime changed from 56s to 33s (on NREL Peregrine
  system)

- Tests on the OWEZ (9x9x1 km) precursor run indicates that the wall-clock time
  per Nalu timestep reduced to approximately 4s/time-step from 9s/time-step with
  the original implementation.